Moved over the comment fix to a new type of config class, so that this is done seemlessly
