+++ src/main/java/net/TheDgtl/Stargate/refactoring/Refactorer.java
@@ -35,9 +35,6 @@ public class Refactorer {
     private FileConfiguration fileConfig;
     private static final Modificator[] RETCONS;
 
-    static private String END_OF_COMMENT = "_endOfComment_";
-    static private String START_OF_COMMENT = "comment_";
-
     static {
         RETCONS = new Modificator[]{
                 new RetCon1_0_0()
@@ -64,81 +61,6 @@ public class Refactorer {
         }
         return config;
     }
-    
-    /**
-     * Reads a file with comments, and recreates them into yaml mappings.
-     * A mapping follows this format: comment_{CommentNumber}: "The comment" 
-     * <p>
-     * This needs to be done as comments otherwise get removed using
-     * the {@link FileConfiguration#save(File)} method. The config
-     * needs to be saved if a config value has changed.
-     * @throws IOException
-     */
-    public void convertCommentsToYAMLMappings() throws IOException {
-        BufferedReader bReader = FileHelper.getBufferedReader(configFile);
-        String line;
-        String newText = "";
-        /*
-         * A list of each stored comment (which is an list of comment lines)
-         * A comment is defined as a set of lines that start with #, this set 
-         * can not contain an empty line.
-         */
-        List<List<String>> comments = new ArrayList<>();
-        int counter = 0;
-        int commentNameCounter = 0;
-        int indent = 0;
-        List<String> currentComment;
-        try {
-            while ((line = bReader.readLine()) != null) {
-                if (line.trim().isEmpty()) {
-                    counter = comments.size(); // a cheesy way to move to the next comment
-                    continue;
-                }
-                if (line.trim().startsWith("#")) {
-                    if (counter >= comments.size()) {
-                        comments.add(new ArrayList<>());
-                    }
-                    currentComment = comments.get(counter);
-                    currentComment.add(line.trim().replaceFirst("#", ""));
-                    continue;
-                }
-                if (!comments.isEmpty()) {
-                    indent = this.countSpaces(line);
-                    for (List<String> comment : comments)
-                        newText = newText + compileCommentMapping(comment, commentNameCounter++, indent);
-                    newText = newText + line + "\n";
-                    comments.clear();
-                    counter = 0;
-                    continue;
-                }
-                newText = newText + line + "\n";
-            }
-        } finally {
-            bReader.close();
-        }
-        
-        File oldFile = new File(configFile.getAbsolutePath() + ".old");
-        if(oldFile.exists())
-            oldFile.delete();
-        configFile.renameTo(oldFile);
-        configFile.createNewFile();
-        BufferedWriter bWriter = FileHelper.getBufferedWriter(configFile);
-        try {
-            bWriter.write(newText);
-        } finally {
-            bWriter.close();
-        }
-        
-    }
-    
-    private String compileCommentMapping(List<String> commentLines, int counter, int indent) {
-        String commentYamlMapping = this.repeat(" ", indent) + START_OF_COMMENT + counter + ": |\n";
-        commentLines.add(this.repeat(" ", indent + 2) + END_OF_COMMENT);
-        for(String commentLine : commentLines) {
-            commentYamlMapping = commentYamlMapping + this.repeat(" ", indent + 2) + commentLine + "\n";
-        }
-        return commentYamlMapping;
-    }
 
     public void insertNewValues(Map<String, Object> config) throws IOException, InvalidConfigurationException {
         fileConfig.load(configFile);
@@ -148,90 +70,7 @@ public class Refactorer {
         fileConfig.set("configVersion", Stargate.CURRENT_CONFIG_VERSION);
         fileConfig.save(configFile);
     }
-
-    public void convertYAMLMappingsToComments() throws IOException, InvalidConfigurationException {
-        BufferedReader bReader = FileHelper.getBufferedReader(configFile);
-        fileConfig.load(configFile);
-
-        String finalText = "";
-        String line;
-        boolean isSkippingComment = false;
-        try {
-            while ((line = bReader.readLine()) != null) {
-                if (isSkippingComment) {
-                    if (line.contains(END_OF_COMMENT))
-                        isSkippingComment = false;
-                    continue;
-                }
-                // TODO: Create a custom method with Java 11's String.strip() if necessary
-                String possibleComment = line.trim();
-                if (possibleComment.startsWith(START_OF_COMMENT)) {
-                    int indent = countSpaces(line);
-                    String lastKeyName = possibleComment.split(":")[0];
-                    Set<String> keys = fileConfig.getKeys(true);
-                    String key = "";
-                    for(String possibleKey : keys) {
-                        if(possibleKey.contains(lastKeyName))
-                            key = possibleKey;
-                    }
-                    String comment = fileConfig.getString(key);
-                    String[] commentLines = comment.split("\n");
-                    line = "";
-                    /*
-                     * Go through every line, except the last one, which is just going to be a
-                     * ENDOFCOMMENT identifier
-                     */
-                    for (int i = 0; i < commentLines.length - 1; i++) {
-                        line = line + "\n" + repeat(" ", indent) + "# " + commentLines[i];
-                    }
-                    isSkippingComment = true;
-                }
-                finalText = finalText + line + "\n";
-            }
-        } finally {
-            bReader.close();
-        }
-
-        BufferedWriter writer = FileHelper.getBufferedWriter(configFile);
-        try {
-            writer.write(finalText);
-        } finally {
-            writer.close();
-        }
-    }
-
-    /**
-     * Repeats the given string the given amount of times
-     * 
-     * @param repeatingString <p>The string to repeat</p>
-     * @param repetitions <p>The number of times to repeat the string</p>
-     * @return <p>The string repeated the given amount of times</p>
-     */
-    private String repeat(String repeatingString, int repetitions) {
-        StringBuilder builder = new StringBuilder();
-        for (int i = 0; i < repetitions; i++) {
-            builder.append(repeatingString);
-        }
-        return builder.toString();
-    }
-
-    /**
-     * Count the spaces at the start of a line. there's probably an already existing
-     * method for this, but meh
-     *
-     * @param line
-     * @return
-     */
-    private int countSpaces(String line) {
-        int spaceAmount = 0;
-        for (char aChar : line.toCharArray()) {
-            if (aChar == ' ')
-                spaceAmount++;
-            else
-                break;
-        }
-        return spaceAmount;
-    }
+    
 
     /**
      * Used in debug, when you want to see the state of the currently stored
